Move full component to one file
