+++ src/components/tokenassistant.js
@@ -1,46 +1,160 @@
+import React, { Component } from 'react';
+import {
+    CLIENT_ID,
+    ISSUER,
+    SCOPE
+} from './constants';
+import axios from 'axios';
 
-import React from 'react'
-import { TokenAssistantContextConsumer } from './context-provider';
+const { Provider, Consumer } = React.createContext();
+
+class TokenAssistantContextProvider extends Component {
+    config = {};
+    count = 0;
+    tokenAssistant;
+    debug = false;
+    state = {
+        isLoggedIn: false,
+    };
+
+    constructor(props) {
+        super(props);
+
+        this.debug = props.debug === "true" ? true : false;
+        this.loadConfiguration();
+    }
+
+    render() {
+        return (
+            <Provider value={{
+                isLoggedIn: this.state.isLoggedIn, userName: this.state.userName, accessToken:
+                    this.state.accessToken, login: this.getToken, logout: this.logout
+            }}>
+                {this.props.children}
+            </Provider>
+        );
+    }
+
+    loadConfiguration = () => {
+        axios.get(ISSUER + "/.well-known/openid-configuration").then(response => {
+            this.config = response.data;
+            console.log(response.data);
+            this.addScriptToIndexFile();
+            this.tryLoadTokenAssistant();
+        });
+    };
+
+
+    tryLoadTokenAssistant = () => {
+        console.log("Loading assistant..");
+        if (window.curity) {
+            if (!this.tokenAssistant) {
+                this.loadTokenAssistant();
+                console.log("DONE");
+            }
+            console.log("Already loaded");
+        }
+        else {
+            setTimeout(() => {
+                this.count++;
+                if (this.count > 100) {
+                    return false;
+                }
+                this.tryLoadTokenAssistant();
+            }, 20);
+        }
+    }
+
+    addScriptToIndexFile = () => {
+        var head = window.document.head;
+        var script = window.document.createElement('script');
+        script.type = 'text/javascript';
+        script.src = this.config.assisted_token_endpoint +
+            '/resources/js/assisted-token.js';
+        script.id = 'assisted-token-js-script';
+        head.appendChild(script);
+    }
+
+    loadTokenAssistant = () => {
+        if (!window.curity) {
+            throw new Error('Assisted token javascript was not found.' +
+                ' Make sure the server is running and/or update URL ' +
+                'of #assisted-token-js-script script');
+        }
+        window.curity.debug = this.debug;
+        this.tokenAssistant = window.curity.token.assistant({
+            clientId: CLIENT_ID,
+        });
+    }
+
+    getToken = () => {
+        if (!this.tokenAssistant) {
+            console.error('Token Assistant is undefined.');
+            return false;
+        }
+        this.tokenAssistant.loginIfRequired({ "scope": SCOPE }).then((msg) => {
+            console.log('Retrieved tokens', msg);
+            let additionalData = this.tokenAssistant.getAdditionalData();
+            console.log('additional data', additionalData)
+
+            this.setState({ isLoggedIn: true, accessToken: msg.token, userName: additionalData.subject });
+        }).fail((err) => {
+            console.log('Failed to retrieve tokens', err);
+        });
+    }
+
+    logout = () => {
+        if (!this.tokenAssistant) {
+            console.error('Token Assistant is undefined.');
+            return false;
+        }
+        this.tokenAssistant.logout().then(() => {
+            this.setState({ accessToken: undefined, userName: undefined, isLoggedIn: false });
+        }).fail((err) => {
+            console.err("Failed to logout.", err);
+        });
+    }
+}
 
 const StateBox = (props) => {
     return (
-        <TokenAssistantContextConsumer>
+        <Consumer>
             {(context) => context.isLoggedIn ? <pre>{context.accessToken}</pre> : ''}
-        </TokenAssistantContextConsumer>
+        </Consumer>
     )
 }
 
 const Greeting = (props) => {
     return (
-        <TokenAssistantContextConsumer>
+        <Consumer>
             {context => context.userName !== undefined ? <h1>Hi {context.userName}</h1> : <h1>Welcome!</h1>}
-        </TokenAssistantContextConsumer>
+        </Consumer>
     )
 }
 
 const ControlPanel = () => {
     return (
-        <TokenAssistantContextConsumer>
+        <Consumer>
             {(context) => {
-                const loginButton = !context.isLoggedIn ?  <button onClick={context.login}>Login</button> : '';
-                const logoutButton = context.isLoggedIn ?  <button onClick={context.logout}>Logout</button> : ''
-                const getTokenButton = context.isLoggedIn ?  <button onClick={context.login}>Get Token</button> : '';
+                const loginButton = !context.isLoggedIn ? <button onClick={context.login}>Login</button> : '';
+                const logoutButton = context.isLoggedIn ? <button onClick={context.logout}>Logout</button> : ''
+                const getTokenButton = context.isLoggedIn ? <button onClick={context.login}>Get Token</button> : '';
                 return (<p>{loginButton} {logoutButton} {getTokenButton}</p>);
             }
             }
-        </TokenAssistantContextConsumer>
+        </Consumer>
     )
 }
 
 function TokenAssistant(props) {
     return (
         <div className="tokenassistant">
-            <Greeting />
-            <ControlPanel />
-            <StateBox />
-        </div>
-    );
+            <TokenAssistantContextProvider debug={props.debug}>
+                <Greeting />
+                <ControlPanel />
+                <StateBox />
+            </TokenAssistantContextProvider>
+        </div>);
 }
-
 
 export default TokenAssistant
