+++ src/pages/index.js
@@ -2,15 +2,12 @@ import React from 'react'
 
 import Layout from "../components/layout"
 import SEO from "../components/seo"
-import { TokenAssistantContextProvider } from '../components/context-provider'
 import TokenAssistant from '../components/tokenassistant'
 
 const IndexPage = () => (
   <Layout>
     <SEO title="Home" />
-    <TokenAssistantContextProvider debug="true">
-      <TokenAssistant />
-    </TokenAssistantContextProvider>
+    <TokenAssistant debug="true"/>
   </Layout>
 )
